application/client/web/grapesjs/src/app/screen297232/screen297232.component.spec.ts
import { async, ComponentFixture, TestBed } from '@angular/core/testing';

import { Screen297232Component } from './screen297232.component';
import { Screen297232Service } from './screen297232.service'
import { of, throwError } from 'rxjs';
import { SharedService } from 'src/shared/shared.service';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { HttpClient } from '@angular/common/http';
import { NgSelectModule } from '@ng-select/ng-select';
import { FormsModule, ReactiveFormsModule } from '@angular/forms';
import { Screen297232Component } from './screen297232.component';


describe('Screen297232Component', () => {
  let component: Screen297232Component;
  let fixture: ComponentFixture<Screen297232Component>;
  let service: Screen297232Service;
  let sharedServiceMock = jasmine.createSpyObj('SharedService', ['methodName1', 'methodName2']);
  let httpClient: HttpClientTestingModule;

  beforeEach(async(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, 
        FormsModule, ReactiveFormsModule,
      ],
      declarations: [ Screen297232Component ],
      providers: [ Screen297232Service, 
        { provide: SharedService, useValue: sharedServiceMock},
      ]
    })
    .compileComponents();
  }));

  beforeEach(() => {
    fixture = TestBed.createComponent(Screen297232Component);
    component = fixture.componentInstance;
    service = TestBed.inject(Screen297232Service);
    httpClient = TestBed.inject(HttpClient);
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  // ngOnInit application onload
  it('should set the created_by property with the value from sessionStorage', () => {
    const mockEmail = 'test@example.com';
    spyOn(sessionStorage, 'getItem').and.returnValue(mockEmail);
    component.ngOnInit();

    expect(sessionStorage.getItem).toHaveBeenCalledWith('email');
    expect(component.grapesjstest.created_by).toEqual(mockEmail);

  });
  


});